init file now runs auto-installer
The init file will now check if you have the right addons, and install only the ones you are missing.
HOWEVER, one of these addons is 37mb, another is 2mb, so if you click New > BLUI, and you're missing Mr Mannequins Tools, and Textools, you'll be left looking at what appears to be a frozen Blender, as the complete install process (6 addons) took a full 5 minutes of waiting on a stalled Blender to come through. 
But come through it did!
So I will have to update the docs about that.
There was also a small formatting error in the keymap that caused the Walk modal map to be deleted! What!?!?
But it has also been fixed.